refactor: restore Qlty baseline
