+++ .agents/scripts/_knowledge_social_outbound_runtime.py
@@ -16,28 +16,14 @@ from _knowledge_social_outbound import (
     _verified_operation,
     now_epoch,
 )
+from _knowledge_social_outbound_claim import ClaimRequest, claim_operation
+from _knowledge_social_outbound_cooldown import (
+    active_connection_cooldown,
+    defer_claim_for_cooldown,
+)
+from _knowledge_social_outbound_queries import due_operation_ids, outbound_health_rows
 from knowledge_social_import import canonical_json
 from knowledge_social_store import SocialStoreError, validate_opaque
-
-CLAIM_OPERATION_SQL = """UPDATE outbound_operations
-                  SET state='claimed',claim_token=?,claimed_by=?,claim_expires_at=?,
-                      last_attempt_id=?,updated_at=?
-                WHERE operation_id=? AND state='approved'"""
-INSERT_ATTEMPT_SQL = """INSERT INTO outbound_attempts(
-                attempt_id,operation_id,claim_token,executor_id,status,started_at)
-               VALUES(?,?,?,?,?,?)"""
-
-
-@dataclass(frozen=True)
-class ClaimRequest:
-    """Owner, executor, and lease values for one atomic operation claim."""
-
-    operation_id: str
-    principal_id: str
-    executor_id: str
-    current_time: int
-    claim_seconds: int
-
 
 @dataclass(frozen=True)
 class AttemptOutcome:
@@ -48,24 +34,6 @@ class AttemptOutcome:
     failure_class: str | None = None
     finished_at: int | None = None
     retry_after: int | None = None
-
-
-def active_connection_cooldown(
-    database: sqlite3.Connection, connection_id: str, current_time: int
-) -> int | None:
-    """Return the latest active reset boundary for one exact connection."""
-    connection_id = validate_opaque(connection_id, "connection_id")
-    if current_time < 0:
-        raise SocialStoreError("cooldown time must be a non-negative epoch")
-    row = database.execute(
-        """SELECT MAX(CAST(retry_after AS INTEGER)) AS reset_at
-             FROM sync_runs
-            WHERE connection_id=? AND status='paused' AND failure_class='rate_limit'
-              AND retry_after!='' AND retry_after NOT GLOB '*[^0-9]*'
-              AND CAST(retry_after AS INTEGER)>?""",
-        (connection_id, current_time),
-    ).fetchone()
-    return int(row["reset_at"]) if row and row["reset_at"] is not None else None
 
 
 def record_provider_checkpoint(
@@ -91,206 +59,6 @@ def record_provider_checkpoint(
     ).rowcount
     if changed != 1:
         raise SocialStoreError("outbound provider checkpoint is stale")
-
-
-def due_operation_ids(
-    database: sqlite3.Connection,
-    principal_id: str,
-    current_time: int,
-    limit: int,
-) -> list[str]:
-    """Return deterministic due IDs with a current exact-intent approval."""
-    if limit < 1 or limit > 100:
-        raise SocialStoreError("due limit must be between 1 and 100")
-    principal_id = validate_opaque(principal_id, "principal_id")
-    rows = database.execute(
-        """SELECT DISTINCT o.* FROM outbound_operations o
-              JOIN outbound_approvals a ON a.operation_id=o.operation_id
-             WHERE o.state='approved' AND o.scheduled_at<=?
-               AND o.created_by=? AND a.principal_id=?
-               AND a.intent_sha256=o.intent_sha256
-               AND a.revoked_at IS NULL AND a.expires_at>?
-               AND NOT EXISTS(
-                   SELECT 1 FROM sync_runs s
-                    WHERE s.connection_id=o.connection_id
-                      AND s.status='paused' AND s.failure_class='rate_limit'
-                      AND s.retry_after!=''
-                      AND s.retry_after NOT GLOB '*[^0-9]*'
-                      AND CAST(s.retry_after AS INTEGER)>?
-               )
-              ORDER BY o.scheduled_at,o.operation_id LIMIT ?""",
-        (
-            current_time,
-            principal_id,
-            principal_id,
-            current_time,
-            current_time,
-            limit,
-        ),
-    ).fetchall()
-    return [
-        str(_verified_operation(database, row["operation_id"])["operation_id"])
-        for row in rows
-    ]
-
-
-def outbound_health_rows(
-    database: sqlite3.Connection,
-    principal_id: str,
-    current_time: int,
-) -> list[dict[str, Any]]:
-    """Return content-free operation evidence for health aggregation."""
-    principal_id = validate_opaque(principal_id, "principal_id")
-    if current_time < 0:
-        raise SocialStoreError("health time must be a non-negative epoch")
-    has_reconciliations = database.execute(
-        "SELECT 1 FROM sqlite_master WHERE type='table' "
-        "AND name='outbound_reconciliations'"
-    ).fetchone()
-    attempt_status = (
-        "COALESCE(r.resolved_state,a.status)" if has_reconciliations else "a.status"
-    )
-    finished_at = (
-        "COALESCE(r.reconciled_at,a.finished_at)"
-        if has_reconciliations
-        else "a.finished_at"
-    )
-    failure_class = (
-        "CASE WHEN r.outcome='succeeded' THEN NULL "
-        "WHEN r.outcome='not-sent' THEN 'reconciled_not_sent' "
-        "ELSE a.failure_class END"
-        if has_reconciliations
-        else "a.failure_class"
-    )
-    reconciliation_join = (
-        "LEFT JOIN outbound_reconciliations r ON r.attempt_id=a.attempt_id"
-        if has_reconciliations
-        else ""
-    )
-    rows = database.execute(
-        f"""SELECT o.operation_id,o.provider,o.connection_id,o.action,o.state,
-                  o.scheduled_at,o.updated_at,o.claim_expires_at,
-                  a.attempt_id,{attempt_status} AS attempt_status,
-                  a.provider_started_at,{finished_at} AS finished_at,
-                  {failure_class} AS failure_class,
-                  CASE WHEN EXISTS(
-                      SELECT 1 FROM outbound_approvals p
-                       WHERE p.operation_id=o.operation_id
-                         AND p.principal_id=o.created_by
-                         AND p.intent_sha256=o.intent_sha256
-                         AND p.revoked_at IS NULL AND p.expires_at>?
-                   ) THEN 1 ELSE 0 END AS has_current_approval
-             FROM outbound_operations o
-              LEFT JOIN outbound_attempts a ON a.attempt_id=o.last_attempt_id
-              {reconciliation_join}
-             WHERE o.created_by=?
-             ORDER BY o.connection_id,o.action,o.created_at,o.operation_id""",  # nosec B608 -- fixed internal projections
-        (current_time, principal_id),
-    ).fetchall()
-    return [dict(row) for row in rows]
-
-
-def _claimable_operation(
-    database: sqlite3.Connection, request: ClaimRequest, principal_id: str
-) -> sqlite3.Row:
-    row = _verified_operation(database, request.operation_id)
-    if (
-        row["state"] != "approved"
-        or int(row["scheduled_at"]) > request.current_time
-        or row["created_by"] != principal_id
-    ):
-        raise SocialStoreError("operation is not due and approved")
-    if active_connection_cooldown(
-        database, str(row["connection_id"]), request.current_time
-    ) is not None:
-        raise SocialStoreError("operation account is in provider cooldown")
-    approval = database.execute(
-        """SELECT approval_id FROM outbound_approvals
-            WHERE operation_id=? AND principal_id=? AND intent_sha256=?
-              AND revoked_at IS NULL AND expires_at>?
-            ORDER BY approved_at DESC LIMIT 1""",
-        (
-            request.operation_id,
-            principal_id,
-            row["intent_sha256"],
-            request.current_time,
-        ),
-    ).fetchone()
-    if approval is None:
-        raise SocialStoreError("operation approval is missing or expired")
-    return row
-
-
-def _persist_claim(
-    database: sqlite3.Connection,
-    request: ClaimRequest,
-    row: sqlite3.Row,
-    executor_id: str,
-) -> tuple[int, str]:
-    claim_token = int(row["claim_token"]) + 1
-    attempt_id = _new_id("att")
-    changed = database.execute(
-        CLAIM_OPERATION_SQL,
-        (
-            claim_token,
-            executor_id,
-            request.current_time + request.claim_seconds,
-            attempt_id,
-            request.current_time,
-            request.operation_id,
-        ),
-    ).rowcount
-    if changed != 1:
-        raise SocialStoreError("operation claim lost a concurrent race")
-    database.execute(
-        INSERT_ATTEMPT_SQL,
-        (
-            attempt_id,
-            request.operation_id,
-            claim_token,
-            executor_id,
-            "running",
-            request.current_time,
-        ),
-    )
-    return claim_token, attempt_id
-
-
-def claim_operation(
-    database: sqlite3.Connection, request: ClaimRequest
-) -> ClaimedOperation:
-    """Atomically claim one due operation and persist its sole running attempt."""
-    if request.claim_seconds < 1 or request.claim_seconds > 3600:
-        raise SocialStoreError("claim_seconds must be between 1 and 3600")
-    principal_id = validate_opaque(request.principal_id, "principal_id")
-    executor_id = validate_opaque(request.executor_id, "executor_id")
-    database.execute("BEGIN IMMEDIATE")
-    try:
-        row = _claimable_operation(database, request, principal_id)
-        claim_token, attempt_id = _persist_claim(
-            database, request, row, executor_id
-        )
-        database.execute("COMMIT")
-    except Exception:
-        if database.in_transaction:
-            database.execute("ROLLBACK")
-        raise
-    return ClaimedOperation(
-        operation_id=request.operation_id,
-        provider=str(row["provider"]),
-        action=str(row["action"]),
-        remote_account_id=str(row["remote_account_id"]),
-        target_remote_id=row["target_remote_id"],
-        destination_remote_id=row["destination_remote_id"],
-        payload=row["payload"],
-        subject=row["subject"],
-        app_profile=row["app_profile"],
-        username=row["username"],
-        claim_token=claim_token,
-        attempt_id=attempt_id,
-        media_path=row["media_path"],
-        media_sha256=row["media_sha256"],
-    )
 
 
 def mark_provider_started(
@@ -342,86 +110,6 @@ def mark_provider_started(
         raise SocialStoreError("outbound provider boundary is stale or already marked")
 
 
-def defer_claim_for_cooldown(
-    database: sqlite3.Connection,
-    claimed: ClaimedOperation,
-    executor_id: str,
-    current_time: int,
-) -> dict[str, Any] | None:
-    """Return a pre-provider claim to approved while preserving its attempt."""
-    executor_id = validate_opaque(executor_id, "executor_id")
-    database.execute("BEGIN IMMEDIATE")
-    try:
-        row = database.execute(
-            """SELECT o.connection_id,o.state,o.claim_token,o.claimed_by,
-                      o.last_attempt_id,a.status,a.provider_started_at
-                 FROM outbound_operations o
-                 JOIN outbound_attempts a ON a.attempt_id=o.last_attempt_id
-                WHERE o.operation_id=?""",
-            (claimed.operation_id,),
-        ).fetchone()
-        expected = (
-            "claimed",
-            claimed.claim_token,
-            executor_id,
-            claimed.attempt_id,
-            "running",
-            None,
-        )
-        actual = tuple(row)[1:] if row is not None else ()
-        if actual != expected:
-            raise SocialStoreError("cooldown deferral claim is stale")
-        reset_at = active_connection_cooldown(
-            database, str(row["connection_id"]), current_time
-        )
-        if reset_at is None:
-            database.execute("ROLLBACK")
-            return None
-        attempt_changed = database.execute(
-            """UPDATE outbound_attempts
-                  SET status='failed',finished_at=?,failure_class='rate_limit',
-                      diagnostics=?
-                WHERE attempt_id=? AND operation_id=? AND claim_token=?
-                  AND executor_id=? AND status='running'
-                  AND provider_started_at IS NULL""",
-            (
-                current_time,
-                canonical_json({"phase": "pre-provider", "reason": "cooldown"}),
-                claimed.attempt_id,
-                claimed.operation_id,
-                claimed.claim_token,
-                executor_id,
-            ),
-        ).rowcount
-        operation_changed = database.execute(
-            """UPDATE outbound_operations
-                  SET state='approved',claimed_by=NULL,claim_expires_at=NULL,updated_at=?
-                WHERE operation_id=? AND state='claimed' AND claim_token=?
-                  AND claimed_by=? AND last_attempt_id=?""",
-            (
-                current_time,
-                claimed.operation_id,
-                claimed.claim_token,
-                executor_id,
-                claimed.attempt_id,
-            ),
-        ).rowcount
-        if attempt_changed != 1 or operation_changed != 1:
-            raise SocialStoreError("cooldown deferral claim is inconsistent")
-        database.execute("COMMIT")
-    except Exception:
-        if database.in_transaction:
-            database.execute("ROLLBACK")
-        raise
-    return {
-        "operation_id": claimed.operation_id,
-        "attempt_id": claimed.attempt_id,
-        "state": "approved",
-        "failure_class": "rate_limit",
-        "retry_after": reset_at,
-    }
-
-
 def _assert_outcome_fields(
     status: str, provider_remote_id: str | None, failure_class: str | None
 ) -> None:
@@ -445,13 +133,12 @@ def _validated_outcome(outcome: AttemptOutcome) -> AttemptOutcome:
     if outcome.failure_class is not None and outcome.failure_class not in FAILURE_CLASSES:
         raise SocialStoreError("invalid outbound failure class")
     retry_after = outcome.retry_after
-    if retry_after is not None and (
-        isinstance(retry_after, bool)
-        or not isinstance(retry_after, int)
-        or retry_after <= finished_at
-        or outcome.failure_class != "rate_limit"
-    ):
-        raise SocialStoreError("invalid outbound rate-limit reset")
+    if retry_after is not None:
+        valid_retry = isinstance(retry_after, int) and not isinstance(retry_after, bool)
+        if not valid_retry or retry_after <= finished_at:
+            raise SocialStoreError("invalid outbound rate-limit reset")
+        if outcome.failure_class != "rate_limit":
+            raise SocialStoreError("invalid outbound rate-limit reset")
     _assert_outcome_fields(
         outcome.status, provider_remote_id, outcome.failure_class
     )
