+++ .agents/scripts/social_provider_health.py
@@ -21,6 +21,29 @@ from _knowledge_social_outbound_runtime import (
     active_connection_cooldown,
     outbound_health_rows,
 )
+from _social_provider_health_actions import (
+    ActionContext,
+    DimensionState,
+    aggregate_dimensions as _aggregate_dimensions,
+    aggregate_status as _aggregate_status,
+    dimensions as _dimensions,
+    fallback as _fallback,
+    next_action as _next_action,
+    read_action_record as _read_action_record,
+    readiness_status as _readiness_status,
+    write_action_record as _write_action_record,
+)
+from _social_provider_health_evidence import (
+    QUEUE_STATES,
+    authentication as _authentication,
+    connections as _connections,
+    evidence as _evidence,
+    freshness as _freshness,
+    json_array as _json_array,
+    latest_sync as _latest_sync,
+    queue as _queue,
+    stale_seconds as _stale_seconds,
+)
 from knowledge_social_registry import (
     PROVIDERS,
     ProviderRegistryError,
@@ -30,31 +53,6 @@ from knowledge_social_store import SCHEMA_VERSION, SocialStoreError
 
 SCHEMA = "aidevops.social-provider-health/v1"
 DEFAULT_STALE_SECONDS = 86_400
-UNWIRED_WRITE_PROVIDERS = frozenset(
-    ("meta_facebook", "meta_instagram", "meta_threads", "tiktok")
-)
-QUEUE_STATES = (
-    "draft",
-    "approved",
-    "claimed",
-    "succeeded",
-    "failed",
-    "unknown",
-    "cancelled",
-)
-DIMENSIONS = (
-    "catalogued",
-    "deployed",
-    "installed",
-    "configured",
-    "enabled",
-    "authenticated",
-    "authorized",
-    "reachable",
-    "runtime_compatible",
-    "tool_visible",
-    "usable",
-)
 CATALOGUED_PROVIDER_IDS = frozenset((*PROVIDERS, *OUTBOUND_PROVIDER_ACTIONS))
 MIN_HEALTH_SCHEMA_VERSION = 7
 
@@ -69,363 +67,11 @@ def require_health_schema(database: sqlite3.Connection) -> None:
         )
 
 
-def _json_object(value: str, label: str) -> dict[str, Any]:
-    try:
-        parsed = json.loads(value)
-    except json.JSONDecodeError as error:
-        raise SocialStoreError(f"stored social {label} is invalid") from error
-    if not isinstance(parsed, dict):
-        raise SocialStoreError(f"stored social {label} must be an object")
-    return parsed
-
-
-def _json_array(value: str, label: str) -> list[str]:
-    try:
-        parsed = json.loads(value)
-    except json.JSONDecodeError as error:
-        raise SocialStoreError(f"stored social {label} is invalid") from error
-    if not isinstance(parsed, list) or any(not isinstance(item, str) for item in parsed):
-        raise SocialStoreError(f"stored social {label} must be an array of text")
-    return sorted(set(parsed))
-
-
-def _stale_seconds(row: sqlite3.Row, fallback: int) -> int:
-    policy = _json_object(str(row["policy_json"]), "health policy")
-    value = policy.get("health_stale_seconds", fallback)
-    if isinstance(value, bool) or not isinstance(value, int) or value < 60:
-        raise SocialStoreError(
-            "stored social health_stale_seconds must be an integer of at least 60"
-        )
-    return value
-
-
-def _connections(database: sqlite3.Connection) -> list[sqlite3.Row]:
-    return database.execute(
-        "SELECT connection_id,provider,auth_profile_ref,enabled_streams,policy_json "
-        "FROM connections ORDER BY provider,connection_id"
-    ).fetchall()
-
-
-def _latest_sync(
-    database: sqlite3.Connection, connection_id: str
-) -> sqlite3.Row | None:
-    return database.execute(
-        """SELECT status,failure_class,retry_after,started_at,completed_at
-             FROM sync_runs WHERE connection_id=?
-             ORDER BY COALESCE(completed_at,started_at,0) DESC,rowid DESC LIMIT 1""",
-        (connection_id,),
-    ).fetchone()
-
-
-def _latest_stream_sync(
-    database: sqlite3.Connection, connection_id: str, stream: str
-) -> sqlite3.Row | None:
-    return database.execute(
-        """SELECT status,failure_class,retry_after,started_at,completed_at
-             FROM sync_runs WHERE connection_id=? AND stream=?
-             ORDER BY COALESCE(completed_at,started_at,0) DESC,rowid DESC LIMIT 1""",
-        (connection_id, stream),
-    ).fetchone()
-
-
 def _canonical_provider(provider: str) -> str:
     try:
         return resolve_provider(provider).provider
     except ProviderRegistryError:
         return provider
-
-
-def _queue(rows: list[dict[str, Any]], now: int) -> dict[str, int]:
-    result = {state: 0 for state in QUEUE_STATES}
-    for row in rows:
-        state = str(row["state"])
-        if state in result:
-            result[state] += 1
-    result["due"] = sum(
-        row["state"] == "approved"
-        and int(row["scheduled_at"]) <= now
-        and bool(row["has_current_approval"])
-        for row in rows
-    )
-    result["leased"] = sum(row["state"] == "claimed" for row in rows)
-    result["total"] = len(rows)
-    return result
-
-
-def _evidence(
-    sync: sqlite3.Row | None, rows: list[dict[str, Any]]
-) -> dict[str, Any]:
-    events: list[tuple[int, str, str | None, bool]] = []
-    successes: list[int] = []
-    failures: list[tuple[int, str]] = []
-    if sync is not None:
-        observed_at = int(sync["completed_at"] or sync["started_at"] or 0)
-        status = str(sync["status"])
-        failure = sync["failure_class"]
-        reached = status in ("complete", "partial") or failure == "rate_limit"
-        if observed_at:
-            events.append((observed_at, status, failure, reached))
-        if status in ("complete", "partial") and observed_at:
-            successes.append(observed_at)
-        if failure is not None and observed_at:
-            failures.append((observed_at, str(failure)))
-    for row in rows:
-        observed_at = int(row["finished_at"] or 0)
-        status = row["attempt_status"]
-        failure = row["failure_class"]
-        reached = status == "succeeded" or (
-            failure == "rate_limit" and row["provider_started_at"] is not None
-        )
-        if status is not None and observed_at:
-            events.append((observed_at, str(status), failure, reached))
-        if status == "succeeded" and observed_at:
-            successes.append(observed_at)
-        if failure is not None and observed_at:
-            failures.append((observed_at, str(failure)))
-    latest = max(events, default=None, key=lambda event: event[0])
-    latest_failure = max(failures, default=None, key=lambda event: event[0])
-    return {
-        "evidence_at": latest[0] if latest else None,
-        "latest_status": latest[1] if latest else None,
-        "latest_failure": latest[2] if latest else None,
-        "latest_reached": latest[3] if latest else None,
-        "last_success_at": max(successes, default=None),
-        "last_failure_class": latest_failure[1] if latest_failure else None,
-    }
-
-
-def _dimensions(
-    *,
-    configured: bool,
-    enabled: bool,
-    authenticated: bool | None,
-    authorized: bool,
-    reachable: bool | None,
-    usable: bool,
-    catalogued: bool = True,
-    deployed: bool = True,
-    installed: bool = True,
-    runtime_compatible: bool = True,
-    tool_visible: bool = True,
-) -> dict[str, bool | None]:
-    return {
-        "catalogued": catalogued,
-        "deployed": deployed,
-        "installed": installed,
-        "configured": configured,
-        "enabled": enabled,
-        "authenticated": authenticated,
-        "authorized": authorized,
-        "reachable": reachable,
-        "runtime_compatible": runtime_compatible,
-        "tool_visible": tool_visible,
-        "usable": usable,
-    }
-
-
-def _aggregate_dimensions(records: list[dict[str, Any]]) -> dict[str, bool | None]:
-    result: dict[str, bool | None] = {}
-    for dimension in DIMENSIONS:
-        values = [record["dimensions"][dimension] for record in records]
-        result[dimension] = (
-            True if True in values else False if False in values else None
-        )
-    return result
-
-
-def _readiness_status(
-    dimensions: dict[str, bool | None],
-    *,
-    ambiguous: bool,
-    cooldown: bool,
-    stale: bool,
-) -> str:
-    if not dimensions["configured"]:
-        return "unconfigured"
-    if dimensions["authenticated"] is False:
-        return "unauthenticated"
-    if ambiguous:
-        return "ambiguous"
-    if cooldown:
-        return "rate_limited"
-    if dimensions["reachable"] is False:
-        return "unreachable"
-    if stale:
-        return "stale"
-    if dimensions["usable"]:
-        return "usable"
-    if dimensions["authenticated"] is None or dimensions["reachable"] is None:
-        return "unknown"
-    if not dimensions["authorized"]:
-        return "awaiting_approval"
-    return "ready"
-
-
-def _next_action(status: str, mode: str = "aggregate") -> str:
-    if status == "usable" and mode == "read":
-        return "collect_enabled_stream"
-    return {
-        "unconfigured": "configure_selected_account",
-        "unauthenticated": "refresh_authentication",
-        "ambiguous": "reconcile_unknown_receipt",
-        "rate_limited": "wait_for_provider_reset",
-        "unreachable": "inspect_provider_availability",
-        "stale": "refresh_provider_health",
-        "usable": "execute_approved_intent",
-        "awaiting_approval": "create_and_approve_intent",
-        "ready": "wait_for_due_operation",
-        "partial": "inspect_account_evidence",
-        "unknown": "run_provider_preflight",
-    }[status]
-
-
-def _fallback(status: str) -> str:
-    if status == "rate_limited":
-        return "wait_without_retry"
-    if status == "ambiguous":
-        return "owner_reconciliation_required"
-    return "gated_no_mutation"
-
-
-def _authentication(
-    configured: bool, evidence: dict[str, Any]
-) -> tuple[bool | None, bool | None]:
-    if not configured:
-        return False, None
-    failure = evidence["latest_failure"]
-    if failure in ("authorization", "identity"):
-        return False, None
-    if evidence["latest_reached"] is True:
-        return True, True
-    if failure == "provider_unavailable":
-        return None, False
-    return None, None
-
-
-def _freshness(
-    evidence: dict[str, Any], now: int, stale_after: int
-) -> tuple[dict[str, int | bool | None], bool]:
-    evidence_at = evidence["evidence_at"]
-    lag = max(0, now - evidence_at) if evidence_at is not None else None
-    stale = lag is not None and lag > stale_after
-    return (
-        {
-            "evidence_at": evidence_at,
-            "lag_seconds": lag,
-            "stale_after_seconds": stale_after,
-            "stale": stale,
-        },
-        stale,
-    )
-
-
-def _read_action_record(
-    database: sqlite3.Connection,
-    connection_id: str,
-    action: str,
-    now: int,
-    configured: bool,
-    cooldown: bool,
-    quota: dict[str, int | bool | None],
-    stale_after: int,
-) -> dict[str, Any]:
-    evidence = _evidence(
-        _latest_stream_sync(database, connection_id, action), []
-    )
-    authenticated, reachable = _authentication(configured, evidence)
-    freshness, stale = _freshness(evidence, now, stale_after)
-    authorized = authenticated is True and reachable is True
-    usable = bool(authorized and not cooldown and not stale)
-    dimensions = _dimensions(
-        configured=configured,
-        enabled=True,
-        authenticated=authenticated,
-        authorized=authorized,
-        reachable=reachable,
-        usable=usable,
-    )
-    status = _readiness_status(
-        dimensions, ambiguous=False, cooldown=cooldown, stale=stale
-    )
-    return {
-        "action": action,
-        "mode": "read",
-        "dimensions": dimensions,
-        "queue": _queue([], now),
-        "quota": dict(quota),
-        "freshness": freshness,
-        "status": status,
-        "fallback": _fallback(status),
-        "next_action": _next_action(status, "read"),
-    }
-
-
-def _write_action_record(
-    provider: str,
-    action: str,
-    rows: list[dict[str, Any]],
-    now: int,
-    configured: bool,
-    enabled: bool,
-    cooldown: bool,
-    quota: dict[str, int | bool | None],
-    stale_after: int,
-) -> dict[str, Any]:
-    action_rows = [row for row in rows if row["action"] == action]
-    queue = _queue(action_rows, now)
-    evidence = _evidence(None, action_rows)
-    authenticated, reachable = _authentication(configured, evidence)
-    if provider in UNWIRED_WRITE_PROVIDERS:
-        reachable = False
-    freshness, stale = _freshness(evidence, now, stale_after)
-    authorized = any(
-        row["state"] in ("approved", "claimed") and row["has_current_approval"]
-        for row in action_rows
-    )
-    usable = bool(
-        queue["due"]
-        and configured
-        and authenticated is True
-        and reachable is True
-        and authorized
-        and not cooldown
-        and not stale
-        and queue["unknown"] == 0
-    )
-    dimensions = _dimensions(
-        configured=configured,
-        enabled=enabled,
-        authenticated=authenticated,
-        authorized=authorized,
-        reachable=reachable,
-        usable=usable,
-    )
-    status = _readiness_status(
-        dimensions,
-        ambiguous=queue["unknown"] > 0,
-        cooldown=cooldown,
-        stale=stale,
-    )
-    return {
-        "action": action,
-        "mode": "write",
-        "dimensions": dimensions,
-        "queue": queue,
-        "quota": dict(quota),
-        "freshness": dict(freshness),
-        "status": status,
-        "fallback": _fallback(status),
-        "next_action": _next_action(status),
-    }
-
-
-def _aggregate_status(records: list[dict[str, Any]]) -> str:
-    statuses = {str(record["status"]) for record in records}
-    if len(statuses) == 1:
-        return next(iter(statuses))
-    if statuses and statuses <= {"ready", "usable"}:
-        return "usable" if "usable" in statuses else "ready"
-    return "partial"
 
 
 def _account_record(
@@ -454,16 +100,20 @@ def _account_record(
         "reset_at": cooldown_until,
         "cooldown": cooldown,
     }
+    action_context = ActionContext(
+        now=now,
+        configured=configured,
+        enabled=enabled,
+        cooldown=cooldown,
+        quota=quota,
+        stale_after=stale_after,
+    )
     actions = [
         _read_action_record(
             database,
             str(row["connection_id"]),
             action,
-            now,
-            configured,
-            cooldown,
-            quota,
-            stale_after,
+            action_context,
         )
         for action in read_actions
     ]
@@ -472,12 +122,7 @@ def _account_record(
             provider,
             action,
             operations,
-            now,
-            configured,
-            enabled,
-            cooldown,
-            quota,
-            stale_after,
+            action_context,
         )
         for action in write_actions
     )
@@ -488,12 +133,14 @@ def _account_record(
     else:
         authenticated, reachable = _authentication(configured, evidence)
         dimensions = _dimensions(
-            configured=configured,
-            enabled=enabled,
-            authenticated=authenticated,
-            authorized=False,
-            reachable=reachable,
-            usable=False,
+            DimensionState(
+                configured=configured,
+                enabled=enabled,
+                authenticated=authenticated,
+                authorized=False,
+                reachable=reachable,
+                usable=False,
+            )
         )
         status = _readiness_status(
             dimensions,
@@ -538,15 +185,17 @@ def _provider_dimensions(
             spec is not None and spec.entrypoint is not None
         )
         return _dimensions(
-            configured=False,
-            enabled=False,
-            authenticated=None,
-            authorized=False,
-            reachable=None,
-            usable=False,
-            deployed=deployed,
-            installed=deployed,
-            tool_visible=deployed,
+            DimensionState(
+                configured=False,
+                enabled=False,
+                authenticated=None,
+                authorized=False,
+                reachable=None,
+                usable=False,
+                deployed=deployed,
+                installed=deployed,
+                tool_visible=deployed,
+            )
         )
     return _aggregate_dimensions(accounts)
 
@@ -641,15 +290,27 @@ def _global_status(providers: list[dict[str, Any]]) -> str:
     return "blocked"
 
 
+def _validate_options(
+    values: Mapping[str, Any], allowed: set[str], function_name: str
+) -> None:
+    unknown = set(values) - allowed
+    if unknown:
+        name = sorted(unknown)[0]
+        raise TypeError(
+            f"{function_name}() got an unexpected keyword argument '{name}'"
+        )
+
+
 def build_health_report(
     database: sqlite3.Connection,
     principal_id: str,
     now_epoch: int,
-    *,
-    stale_seconds: int = DEFAULT_STALE_SECONDS,
-    provider: str | None = None,
+    **options: Any,
 ) -> dict[str, Any]:
     """Build one deterministic report from existing local content-free evidence."""
+    _validate_options(options, {"stale_seconds", "provider"}, "build_health_report")
+    stale_seconds = options.get("stale_seconds", DEFAULT_STALE_SECONDS)
+    provider = options.get("provider")
     if now_epoch < 0:
         raise SocialStoreError("health time must be a non-negative epoch")
     if stale_seconds < 60:
@@ -734,13 +395,18 @@ def reconcile_provider_receipts(
     database: sqlite3.Connection,
     principal_id: str,
     now_epoch: int,
-    *,
-    decisions: Iterable[ReconciliationRequest] = (),
-    cooldowns: Mapping[str, int] | None = None,
-    limit: int = 10,
-    per_provider_limit: int = 3,
+    **options: Any,
 ) -> dict[str, Any]:
     """Run one bounded local reconciliation pass without provider mutation."""
+    _validate_options(
+        options,
+        {"decisions", "cooldowns", "limit", "per_provider_limit"},
+        "reconcile_provider_receipts",
+    )
+    decisions = options.get("decisions", ())
+    cooldowns = options.get("cooldowns")
+    limit = options.get("limit", 10)
+    per_provider_limit = options.get("per_provider_limit", 3)
     active_cooldowns = (
         connection_cooldowns(database, now_epoch) if cooldowns is None else cooldowns
     )
